add zookeeper read write lock
